colab_requirements.txt
captum
lightning
neptune
optuna
optuna-integration[pytorch_lightning]
python-dotenv
timm
torchinfo
